Clarify checking out code
Rewrite to describe the goal rather than specifying using Git.
Preface specific advice about Git with "if you are using Git".
Clarify that the directory you push from should contain *all* the app
code and config files (assuming it's not worth a digression to explain
how you can configure to look in different folders).

#136104121

When I’m deploying an app to the PaaS for the first time
I need to understand how my files must be checked out of a local folder
So that I can successfully deploy to the PaaS.
Acceptance Criteria
Documentation is clear what is meant by ‘check out’. If there’s a
specific way that users must do this then we should say so.
User Research
During lab testing, a participant used SVN instead of Git to check out
the app. This meant that his manifest file was not in the same
directory as the rest of the app so his push failed. It wasn't clear to
him from the documentation what he should have done differently