Document CRLF issue
Explain to Windows users that pushing with CRLF line breaks can cause
errors.

Story: https://www.pivotaltracker.com/story/show/135421943

##What

Note that Cloud foundry at this time runs all applications in Linux
containers.
Add some instructions to the docs so that new Windows users know that
files with CRLF line endings won't always work

##Suggestions to consider

turn off the autocrlf setting in git
Or make sure that you specify the interpeter to be used in a command in
manifest.yml or a Procfile eg. command: ruby foo.rb

##Why

We had an issue in user research where a user tried to push our sample
ruby app from a windows machine in a git repo with autocrlf turned on.
It failed because Linux was unable to understand the shebang line.