Reorganise Jenkins doc, document rate limit
Reorganise Jenkins docs, try to remove impression that using custom
script deployment is less secure. Document issue with login rate limit

https://www.pivotaltracker.com/story/show/136104045

"In our current jenkins docs
(https://github.com/alphagov/paas-tech-docs/blob/01a111ddb7ea95951e2f0d2
c809b317acc292fd5/source/documentation/deploying_apps/jenkins.md), we
recommend users log in every time job runs. This is so that they don't
have to handle secure storage of CF secrets (which is a file with a
logged-in token cf CLI produces after logging in). However our platform
has protection to lock accounts that log in 5 times in 5 minutes. If
someone has a lot of jenkins jobs, or small jobs that run often, they
can easily hit this limit and lock themselves out. As it has already
happened once and support request had to be risen to analyse and
resolve the situation.
We should either document this (that there is a lockout and what
criteria trigger it) or instead think about secure handling of
logged-in credentials and recommend that to users instead."