Dockerfile for cf deploy image
Contains the cf cli, python and ruby environments for building, testing
and deploying to cf.

It is based on the ruby image to avoid any issue with the ruby
environment (bundler, etc).

The dev packages are used to build gem and python pakage native
extensions. They are removed in the same step to save ~ 150MB.