Remove old URLs from README
Replace all the content about old URLs with the new URL.

The new `docs.cloud` URL was made available in Pivotal story #136353673. The
old `government-paas-developer-docs.readthedocs` URL was redirect to
`docs.cloud` in Pivotal story #136429263.

There is still an outstanding Pivotal story #139284061 to redirect
`paas-tech-docs.cloudapps` to `docs.cloud`, but I think it's sufficient to
remove the notice from the README for now.