Add pull request template
Copy the PaaS team's standard pull request template from alphagov/paas-cf.
This will pre-populate the description for new pull requests. The headings
are designed to help people review pull requests, which is more pertinent
for this repository because the ownership crosses team boundaries and it's
not always clear who should review a change and how.