Fix list formatting in SSH tunnel docs
The extra newline was causing the nested unordered list to be rendered as a
codeblock with literal `+` characters before each line and the next ordered
list item to be rendered as `1.` instead of `2.`.

I'm not happy with the subtly of this change, but no amount of indentation
changes to the unordered list made it render correctly.