Mark snapshot restore as experimental
We think the typical use case for snapshot restore is to create a copy of the
db in a test/staging environment. We know that people who are likely to use it
will use SSH tunnels and run into the problems described in cfc7ec0.

We'd also like feedback from people who try to use it.