Update SSH warning now that the 2GB bug has been fixed
The bug that this relates to has been fixed upstream with version 6.26
of the CLI. This therefore updates our warning to reflect this.

I've left the warning in place but made it specific to 6.24 and 6.25 of
the CLI because there are likely to be many tenants who still have these
versions installed.