MySQL: Some libraries don't use TLS by default
We have found that Rails, Django and PHP all don't use TLS by default. It's
not possible to provide a database URI/URL with query parameters that
satisfies most of them. We're going to defer writing more specific
documentation for each individual language until we understand what people
need a bit better. This might result in more support requests in the
meantime.