Explain that CI accounts need email addresses
It's not currently made obvious that we have a policy that usernames should be
email addresses. One result of this is that users request CI accounts without
including email addresses, the support person has to ask them for an email
address and there's a delay in getting what they need.

This commit makes it clearer that an email address is required when requesting
CI users.