Add sections for PostgreSQL, MySQL and MongoDB
Have added small sections on each backing service with associated h2 headings. This is to call out the current backing services in the left-hand navigation so that users can scan and quickly see these sections. I have not added them as h3 headings as this would mean editing the config yaml file so that all h3 headings are in the left-hand nav which would not be a good experience for the user.