Add setting of DEPS_DIR to ssh profile instructions.
The buildpack launcher process that starts the application sets this
environment variable[1] before sourcinf the profile files as part of
launching the app. The diego ssh handler doesn't set this. This means
that for buildpacks that don't override this in their own .profile files
the environment won't work. The Python buildpack is an example, and
means that it's somewhat tricky to run tasks against the app (eg
Django's manage.py).

[1]https://github.com/cloudfoundry/buildpackapplifecycle/blob/master/launcher/main.go#L37-L40