logit logstash config: clearer parse failure tag
`_grokparsefailure` is a generic tag that any grok filter will throw
by default.  By setting a tag specific to the particular filter, it
makes it much clearer exactly where in the filter processing the parse
failure happened.

This filter configuration already did this for the router grok
parsing (with `_routerparsefailure`); this commit does the same for
the syslog grok parsing.