Update Logstash filter to create fields for app/space/org
We found syslog_host contained these fields but it caused some confusion as it
isn't the most obvious place to find the app name.

Creating separate fields for these will hopefully be more intuitive and make it
easier to produce queries when dealing with multiple apps shipping logs to a
single ELK stack.