Remove file extensions from product page links
Links to the product pages shouldn't have a file extension because they are
rendered without one and that's how they are referenced in the product pages
themselves.

Most of these are in the header, but there were also some references to the
support page within the custom domain documentation.

These also caused the requests to redirect off of `service.gov.uk` and go to
`cloudapps.digital` because of a bug that we're fixing in a separate story.