Use the launcher command to setup the SSH environment
This avoids the need to keep these instructions up to date with the
launcher itself.

According to the usage output[1], the 3rd argument to the launcher is
"metadata". I'm not clear what that is. From looking at the source[2],
it doesn't appear to be used at all at the moment. It's possible a
legacy thing.

[1]Usage: /tmp/lifecycle/launcher <app-directory> <start-command> <metadata>
[2]https://github.com/cloudfoundry/buildpackapplifecycle/blob/ef6149f/launcher/main.go#L16