Add shared responsibility model section
We want to document the responsibility model for the users using Docker
on CF vs. using buildpacks. We can write a document inspired on
https://aws.amazon.com/compliance/shared-responsibility-model/,
including a table or image that would list the different layers
of the application, runtime and platform.

We also add links when explaining buildpacks and docker images.

Signed-off-by: Jonathan Glassman <jonathan.glassman@digital.cabinet-office.gov.uk>