Upgrading plans with same encryption type
Added in content to state that you can only upgrade your PostgreSQL and MySQl plans using the `cf update-services` command if the new plan has the same encryption type as the current plan. If the encryption types are different then you have to create a new service and migrate the data over.