Expand linkchecker testing to catch more 404s
When running linkchecker directly against the build directory, it checks
links that move up the directory hierarchy only for link correctness, not
existence. This means that, before this commit, our standard "make test"
pre-merge only checked that internal links reached pages which exist in
around 40% of cases. Using `--check-extern` isn't an option, as this not
only makes linkchecker check off-site hosts, but will also fail with
`/`-prefixed, relative (internal) links when it discovers that, on the
filesystem, `/stylesheets/foo.css` doesn't exist. There doesn't appear
to be a way to tell linkchecker to assume a file://path/to/build/dir
base URL.

To get round this, we use the fact that linkchecker *will* check the
internal consistency of links when given an http:// URL to check. This
is used in a new Make target: `test-local-http`.

To serve the build dir over HTTP, a simple ruby one-liner was
arbitrarily chosen as "probably having its dependencies present in all
environments which will run this". This might need revisiting if this
assumption turns out to be incorrect. A bit of shell hackery is
needed to clean up the http server and still exit the target with the
linkchecker's exit code.

The existing `test` Make target is redefined to run the fast test target
and then the longer one, allowing devs to have fail-fast behaviour
whilst requiring no changes in CI to pick up the wider set of 404
assertions.