Refer to different domains depending region
Added content stating that app domains are different depending on
which region your account is created in.

We stop mentioning one of the specific domains (previously Ireland
with `.cloudapps.digital`), but instead pointing to a shared
page that explains the different domains for each region.

In the case of examples, we change the examples to point to London
only, as it is the default region.