Workaround tech-docs-gem bug by using HTML headings
There is a bug [1] in tech-docs-gem that causes identical headings on different
pages to be generated with incorrect identifiers. The behaviour is correct for
headings on the same page, but not for different pages as here. Jon Hallam, Jon
Glassman, and I agreed that a workaround using HTML headings was suitable here.

[1] https://github.com/alphagov/tech-docs-template/issues/180