Document how to upgrade from cflinuxfs2
This documentation is a quick explanation of what the upgrade is and
how tenants can manually perform it. The instructions are deliberately
generic because tenants deploy their apps in lots of different ways.
The link to contact our support will help people who feel stuck.

This new content will be linked to [1] by our admin panel.

[1] https://github.com/alphagov/paas-admin/pull/205