Correct reference to route service in private apps
Previously, we pointed readers to the route services documentation to find out 
how to implement an IP whitelist in front of their apps. However, we don't
actually document that or support it as part of the platform. We actually
document how to implement basic auth.

Also adds a pointer to the `re-paas-ip-safelist-service`[1] app for 
implementing an IP whitelist.

[1] https://github.com/alphagov/re-paas-ip-safelist-service