Separate Monitoring docs for apps and services
This commit adds a new entry to the Tech Docs sidebar entitled
"Monitoring backing services". This page describes how to access
metrics and logs for tenants to understand the state of their backing
services.

This content is all moved over from the pre-existing "Monitoring apps"
documentation. While there may be some broken anchor links, I am not
clear that our Tech Docs provide any options for addressing that. To
try and mitigate user confusion I've ensured that both sections of
documentation link to each other.

I have done this because the "Monitoring apps" section was unwieldy
and disordered. It contains a description of every backing service
metric shown by `paas-admin`, and that's way too much content to have
inline. For now I'm not sure where else to put it.

In the next commit I will be adding new content describing how to get
backing service metrics exported to Prometheus. This split should give
that content a good place to live.