Cookie banner and cookie consent functions
Implementation of gtag.js Google analytics and the cookie banner
consent model we have on product pages and admin
(https://github.com/alphagov/paas-admin/pull/1123)

As the parent (tech-docs gem) doesn't have a slot for a cookie banner
we're injecting markup with javascript.
The built-in GA tracking options are also not suitable for us, so we're
implementing the same tracking we have on our other estates.