Remove duplicate link in header
There's already a link to / in the header (The main "GOV.UK  Platform as a Service" link).

Having the same link in the header twice was flagged as an accessibility issue in a recent audit of the tech docs template. This can't be fixed in the gem, but we can fix it in our docs.